Changes for non-AsyncDNS and station IP address reporting.
Added call to Loop while waiting for connection.  loop calls safeLoop that conditionally calls dnsServer->processNextRequest() if not using ESPAsyncDNS.

Moved change back to Station Only mode after connect so InfoPage can contain the actual IP assigned to the station by the connected AccessPoint.  Added additional delays waiting for the infoPage request from the Browser and some additional time on the Browser Refresh to give the station time to actually connect.
This could all probably be done more synchronously with a scripted XMLHttpRequest in place of the current delay and hope method waiting for a connection.